video delete from aws s3 bucket (#13)

+++ src/services/video.service.js
@@ -1,5 +1,6 @@
 import BaseService from './base.service';
 import Video from '../models/video.model';
+import AWS from 'aws-sdk';
 import { transaction } from 'objection';
 
 class VideoService extends BaseService {
@@ -40,22 +41,61 @@ class VideoService extends BaseService {
     }
   }
 
+  async deleteVideosFromS3(videoArray) {
+    var deleteParam = {
+      Bucket: 'media.flexapp.co.uk',
+      Delete: {
+        Objects: videoArray,
+      },
+    };
+    const options = {
+      signatureVersion: 'v4',
+      region: 'eu-west-2',
+    };
+
+    const s3 = new AWS.S3(options);
+
+    s3.deleteObjects(deleteParam, function (err, data) {
+      // error
+      if (err) return false;
+      // deleted
+      else return true;
+    });
+  }
+
+  getVideoObject(videos) {
+    const videoObjArray = [];
+    videos.forEach((video) => {
+      videoObjArray.push({ Key: video.video });
+    });
+    return videoObjArray;
+  }
+
   async deleteVideo(id) {
     const video = await this.findById(id);
-
+    await this.deleteVideosFromS3(this.getVideoObject([video]));
     await Video.query().deleteById(id);
-
     return video;
   }
 
   async deleteVideoByProjectId(projectId) {
-    await Video.query().delete().where('projectId', projectId);
-    return true;
+    const videos = await Video.query().where('projectId', projectId);
+    if (videos.length) {
+      await this.deleteVideosFromS3(this.getVideoObject(videos));
+      const result = await Video.query().delete().where('projectId', projectId);
+      return result;
+    }
+    return false;
   }
 
   async deleteVideoByTopicId(topicId) {
-    await Video.query().delete().where('topicId', topicId);
-    return true;
+    const videos = await Video.query().where('topicId', topicId);
+    if (videos.length) {
+      await this.deleteVideosFromS3(this.getVideoObject(videos));
+      const result = await Video.query().delete().where('topicId', topicId);
+      return result;
+    }
+    return false;
   }
 
   async findByProject(projectId, orderBy = {}) {
@@ -71,7 +111,7 @@ class VideoService extends BaseService {
   }
 
   async findByTopics(topicIds, orderBy = {}) {
-    const { field = '', direction = 'asc' } = orderBy;
+    const { field = 'updatedAt', direction = 'desc' } = orderBy;
 
     let query = Video.query()
       .select('videos.*', 'users.firstName', 'users.lastName', 'users.email')
@@ -85,13 +125,14 @@ class VideoService extends BaseService {
     return query;
   }
 
-  async findByTopic(topicId, orderBy = {}) {
-    const { field = '', direction = 'asc' } = orderBy;
+  async findByTopic(projectId, topicId, orderBy = {}) {
+    const { field = 'updatedAt', direction = 'desc' } = orderBy;
 
     let query = Video.query()
       .select('videos.*', 'users.firstName', 'users.lastName', 'users.email')
       .leftJoin('users', 'videos.userId', 'users.id')
-      .where('topicId', topicId);
+      .where('topicId', topicId)
+      .where('projectId', projectId);
 
     if (field && query) {
       query = query.orderBy(field, direction);
